move UpdateSyncingState into chain.SetTip, so syncing state will be updated during both startup and BlockConnected
